+++ lib/src/controls/inputs/checkbox.dart
@@ -1,6 +1,6 @@
-import 'package:fluent_ui/fluent_ui.dart';
-import 'package:fluentui_system_icons/fluentui_system_icons.dart';
 import 'package:flutter/rendering.dart';
+import 'package:fluent_ui/fluent_ui.dart';
+import 'package:fluentui_icons/fluentui_icons.dart';
 
 class Checkbox extends StatelessWidget {
   const Checkbox({
@@ -25,30 +25,42 @@ class Checkbox extends StatelessWidget {
   Widget build(BuildContext context) {
     debugCheckHasFluentTheme(context);
     final style = context.theme.checkboxStyle.copyWith(this.style);
+    final double size = 22;
     return HoverButton(
       semanticsLabel: semanticsLabel,
       margin: style.margin,
       focusNode: focusNode,
       cursor: (_, state) => style.cursor?.call(state),
-      onPressed: onChanged == null ? null : () => onChanged(!checked),
+      onPressed: onChanged == null
+          ? null
+          : () => onChanged(checked == null ? null : !checked),
       builder: (context, state) {
         return AnimatedContainer(
+          alignment: Alignment.center,
           duration: style.animationDuration,
           curve: style.animationCurve,
           padding: style.padding,
-          decoration: checked
-              ? style.checkedDecoration(state)
-              : style.uncheckedDecoration(state),
-          child: AnimatedSwitcher(
-            duration: style.animationDuration,
-            switchInCurve: style.animationCurve,
-            child: Icon(
-              style.icon,
-              color: checked
-                  ? style.checkedIconColor(state)
-                  : style.uncheckedIconColor(state),
-              key: ValueKey(state),
-            ),
+          height: size,
+          width: size,
+          decoration: () {
+            if (checked == null)
+              return style.thirdstateDecoration(state);
+            else if (checked)
+              return style.checkedDecoration(state);
+            else
+              return style.uncheckedDecoration(state);
+          }(),
+          child: Icon(
+            style.icon,
+            size: 18,
+            color: () {
+              if (checked == null)
+                return style.thirdstateIconColor(state);
+              else if (checked)
+                return style.checkedIconColor(state);
+              else
+                return style.uncheckedIconColor(state);
+            }(),
           ),
         );
       },
@@ -90,79 +102,44 @@ class CheckboxStyle {
   });
 
   static CheckboxStyle defaultTheme(Style style, [Brightness brightness]) {
-    final disabledColor = Colors.grey[100].withOpacity(0.6);
     final radius = BorderRadius.circular(3);
-    final def = CheckboxStyle(
+    final Color accent = style.accentColor ?? Colors.blue;
+    final Color unselected = () {
+      if (brightness == null || brightness == Brightness.light)
+        return Colors.black;
+      return Colors.white;
+    }();
+    return CheckboxStyle(
       cursor: buttonCursor,
-      checkedDecoration: (state) {
-        Color color;
-        if (state.isDisabled)
-          color = disabledColor;
-        else if (state.isHovering || state.isPressing)
-          color = Colors.blue[60];
-        else
-          color = Colors.blue[50];
-        return BoxDecoration(
-          borderRadius: radius,
-          color: color,
-          border: Border.all(style: BorderStyle.none),
-        );
+      checkedDecoration: (state) => BoxDecoration(
+        borderRadius: radius,
+        color: inputColor(accent, state),
+        border: Border.all(style: BorderStyle.none),
+      ),
+      uncheckedDecoration: (state) => BoxDecoration(
+        border: Border.all(
+          width: 0.6,
+          color: state.isDisabled ? kDefaultButtonDisabledColor : unselected,
+        ),
+        color: Colors.transparent,
+        borderRadius: radius,
+      ),
+      thirdstateDecoration: (state) => BoxDecoration(
+        borderRadius: radius,
+        border: Border.all(width: 6.5, color: inputColor(accent, state)),
+      ),
+      checkedIconColor: (_) => Colors.white,
+      uncheckedIconColor: (state) {
+        if (state.isHovering || state.isPressing)
+          return unselected.withOpacity(0.8);
+        return Colors.transparent;
       },
-      thirdstateDecoration: (state) {
-        Color color;
-        if (state.isDisabled)
-          color = disabledColor;
-        else if (state.isHovering || state.isPressing)
-          color = Colors.blue[60];
-        else
-          color = Colors.blue[50];
-        return BoxDecoration(
-          borderRadius: radius,
-          color: color,
-          border: Border.all(
-            width: 4.5,
-            color: Colors.blue,
-          ),
-        );
-      },
-      padding: EdgeInsets.all(4),
+      thirdstateIconColor: (_) => Colors.transparent,
       margin: EdgeInsets.all(4),
-      icon: FluentIcons.checkbox_checked_24_filled,
+      icon: FluentSystemIcons.ic_fluent_checkmark_regular,
       animationDuration: Duration(milliseconds: 200),
       animationCurve: Curves.linear,
     );
-    if (brightness == null || brightness == Brightness.light)
-      return def.copyWith(CheckboxStyle(
-        uncheckedDecoration: (state) => BoxDecoration(
-          border: Border.all(
-            width: 0.6,
-            color: state.isDisabled ? disabledColor : Colors.black,
-          ),
-          color: Colors.transparent,
-          borderRadius: radius,
-        ),
-        checkedIconColor: (_) => Colors.white,
-        uncheckedIconColor: (state) {
-          if (state.isHovering || state.isPressing) return Colors.black;
-          return Colors.transparent;
-        },
-      ));
-    else
-      return def.copyWith(CheckboxStyle(
-        uncheckedDecoration: (state) => BoxDecoration(
-          border: Border.all(
-            width: 0.6,
-            color: state.isDisabled ? disabledColor : Colors.white,
-          ),
-          color: Colors.transparent,
-          borderRadius: radius,
-        ),
-        checkedIconColor: (_) => Colors.white,
-        uncheckedIconColor: (state) {
-          if (state.isHovering || state.isPressing) return Colors.white;
-          return Colors.transparent;
-        },
-      ));
   }
 
   CheckboxStyle copyWith(CheckboxStyle style) {
@@ -178,6 +155,8 @@ class CheckboxStyle {
       animationDuration: style?.animationDuration ?? animationDuration,
       checkedDecoration: style?.checkedDecoration ?? checkedDecoration,
       uncheckedDecoration: style?.uncheckedDecoration ?? uncheckedDecoration,
+      thirdstateDecoration: style?.thirdstateDecoration ?? thirdstateDecoration,
+      thirdstateIconColor: style?.thirdstateIconColor ?? thirdstateIconColor,
     );
   }
 }